+++ src/core/SceneRegistry.ts
@@ -1,7 +1,10 @@
 import type Phaser from 'phaser';
 
 import { BootScene } from '@scenes/BootScene';
+import { EndingScene } from '@scenes/EndingScene';
 import { Map1Scene, Map2Scene, Map3Scene, Map4Scene } from '@scenes/FinalLevelScene';
+import { MainMenuScene } from '@scenes/MainMenuScene';
+import { PauseMenuScene } from '@scenes/PauseMenuScene';
 import { ProductionScene } from '@scenes/ProductionScene';
 import { SceneKey } from '@scenes/SceneKey';
 
@@ -14,6 +17,18 @@ export const sceneRegistry = [
   {
     key: SceneKey.Boot,
     scene: BootScene
+  },
+  {
+    key: SceneKey.MainMenu,
+    scene: MainMenuScene
+  },
+  {
+    key: SceneKey.PauseMenu,
+    scene: PauseMenuScene
+  },
+  {
+    key: SceneKey.Ending,
+    scene: EndingScene
   },
   {
     key: SceneKey.Production,